Refactor: Game files moved to Core/GameDef namespace.
