+++ ShapeEngine/Core/GameDef/GameVirtual.cs
@@ -4,7 +4,7 @@ using ShapeEngine.Geometry.RectDef;
 using ShapeEngine.Input;
 using ShapeEngine.Screen;
 
-namespace ShapeEngine.Core;
+namespace ShapeEngine.Core.GameDef;
 
 public partial class Game
 {
