+++ ShapeEngine/Geometry/SegmentsDef/Segments.cs
@@ -4,6 +4,7 @@ using ShapeEngine.Geometry.PointsDef;
 using ShapeEngine.Geometry.SegmentDef;
 using ShapeEngine.Random;
 using ShapeEngine.StaticLib;
+using Game = ShapeEngine.Core.GameDef.Game;
 
 namespace ShapeEngine.Geometry.SegmentsDef;
